implement ability to sum numbers with decimals

+++ main.c
@@ -3,14 +3,14 @@
 
 int main()
 {
-    int num1;
-    int num2;
+    double num1;
+    double num2;
     printf("Enter first number: ");
-    scanf("%d", &num1);
+    scanf("%lf", &num1);
     printf("Enter second number: ");
-    scanf("%d", &num2);
+    scanf("%lf", &num2);
 
-    printf("\nYour answer is: %d\n", num1 + num2);
+    printf("\nYour answer is: %f\n", num1 + num2);
 
     return 0;
 }
